+++ viewmodel-koin/src/commonTest/kotlin/com/hoc081098/kmp/viewmodel/koin/test/TestViewModelWithExtras.kt
@@ -1,0 +1,8 @@
+package com.hoc081098.kmp.viewmodel.koin.test
+
+import com.hoc081098.kmp.viewmodel.CreationExtras
+import com.hoc081098.kmp.viewmodel.ViewModel
+
+class TestViewModelWithExtras(
+  val extras: CreationExtras,
+) : ViewModel()