+++ viewmodel-koin/src/commonTest/kotlin/com/hoc081098/kmp/viewmodel/koin/test/TestViewModelWithParams.kt
@@ -1,0 +1,8 @@
+package com.hoc081098.kmp.viewmodel.koin.test
+
+import com.hoc081098.kmp.viewmodel.ViewModel
+
+internal class TestViewModelWithParams(
+  val int: Int,
+  val string: String,
+) : ViewModel()