+++ viewmodel/src/androidMain/kotlin/com/hoc081098/kmp/viewmodel/ViewModelStoreOwner.android.kt
@@ -7,12 +7,16 @@ private class AndroidXToKmpViewModelStoreOwner(
   @JvmField val platform: AndroidXViewModelStoreOwner,
 ) : ViewModelStoreOwner {
   override val viewModelStore: ViewModelStore = platform.viewModelStore.toKmp()
+
+  override fun toString(): String = "AndroidXToKmpViewModelStoreOwner(platform=$platform)"
 }
 
 private class KmpToAndroidXViewModelStoreOwner(
   @JvmField val kmp: ViewModelStoreOwner,
 ) : AndroidXViewModelStoreOwner {
   override val viewModelStore: AndroidXViewModelStore = kmp.viewModelStore.toAndroidX()
+
+  override fun toString(): String = "KmpToAndroidXViewModelStoreOwner(kmp=$kmp)"
 }
 
 public fun AndroidXViewModelStoreOwner.toKmp(): ViewModelStoreOwner =