+++ CHANGELOG.md
@@ -10,6 +10,26 @@
   `ViewModel`'s `addCloseable()` now immediately closes the `Closeable` if the `ViewModel` has already received a call to `onCleared()`.
   Now, this behavior is the same for all targets.
 
+### `kmp-viewmodel-koin`
+
+- **Fix** `koinViewModelFactory`: `CreationExtras` passed to `ViewModelFactory.create` will now be
+  passed to the constructor of the ViewModel if it's requested.
+
+  ```kotlin
+  class MyViewModel(val extras: CreationExtras) : ViewModel()
+
+  val myModule: Module = module {
+    factoryOf(::MyViewModel)
+  }
+
+  val factory = koinViewModelFactory<MyViewModel>(
+    scope = KoinPlatformTools.defaultContext().get().scopeRegistry.rootScope,
+  )
+  val extras = buildCreationExtras { /* ... */ }
+  val viewModel: MyViewModel = factory.create(extras)
+  viewModel.extras === extras // true <--- `extras` is the same as `extras` passed to `factory.create(extras)`
+  ```
+
 ## [0.6.2] - Feb 5, 2024
 
 ### Update dependencies